speed controller feature implemented but turning and reverse does not work

+++ STM/Core/Inc/arcade_controller.h
@@ -9,10 +9,9 @@
 #define INC_ARCADE_CONTROLLER_H_
 
 #include "main.h"
-#include "generic_data_type.h"
 #include "motor_pwm_interface.h"
+#include "generic_data_type.h"
 
-
-std_return_type arcade_drive(int8_t rotate, int8_t vertical_speed);
+std_return_type arcade_drive(int rotate, int vertical_speed);
 
 #endif /* INC_ARCADE_CONTROLLER_H_ */
